Correcciones y endpoints para todas las entidades asociadas a la docencia, docente, asignaturas, grado, grupo, falta por confirmar duda para imparte

+++ app_horarios/docencia/models.py
@@ -16,12 +16,12 @@ class Asignaturas(models.Model):
     idasignatura = models.CharField(db_column='IDASIGNATURA', primary_key=True, max_length=10)  # Field name made lowercase.
     nombre = models.CharField(db_column='NOMBRE', max_length=30)  # Field name made lowercase.
     abreviatura = models.CharField(db_column='ABREVIATURA', max_length=20, blank=True, null=True)  # Field name made lowercase.
-    curso_grado = models.DecimalField(db_column='CURSO_GRADO', max_digits=1, decimal_places=0)  # Field name made lowercase.
-    semestre_academico = models.DecimalField(db_column='SEMESTRE_ACADEMICO', max_digits=1, decimal_places=0)  # Field name made lowercase.
-    ects = models.DecimalField(db_column='ECTS', max_digits=2, decimal_places=0)  # Field name made lowercase.
+    curso_grado = models.IntegerField(db_column='CURSO_GRADO')  # Field name made lowercase.
+    semestre_academico = models.IntegerField(db_column='SEMESTRE_ACADEMICO')  # Field name made lowercase.
+    ects = models.IntegerField(db_column='ECTS')  # Field name made lowercase.
     grado_id = models.ForeignKey(Grado, models.CASCADE, db_column='GRADO_ID', max_length=10)  # Field name made lowercase.
     tipo = models.CharField(db_column='TIPO', max_length=1, blank=True, null=True)  # Field name made lowercase.
-    horas_practicas = models.DecimalField(db_column='HORAS_PRACTICAS', max_digits=1, decimal_places=0, blank=True, null=True)  # Field name made lowercase.
+    horas_practicas = models.IntegerField(db_column='HORAS_PRACTICAS', blank=True, null=True)  # Field name made lowercase.
     docentes_imparten = models.ManyToManyField('Docente', through='Imparte', related_name='asignaturas')
 
     class Meta:
